Restructure RobotMechanism
